convert weights to integer

+++ app/src/main/java/io/agritrack/fishtrack/data/dto/FishingHarvestBinDTO.java
@@ -7,7 +7,7 @@ public class FishingHarvestBinDTO {
     public Long id;
     public String fishingId;
     public String harvestBinRFId;
-    public Double quantity;
+    public Integer quantity;
     public String temperature;
 
     public static FishingHarvestBin convert(FishingHarvestBinDTO fishingHarvestBinDTO) {
